# API_SPECIFICATION.yaml
openapi: 3.0.0
info:
  title: Card Clash 21 API
  version: 1.0.0
  description: REST API and WebSocket specification for Card Clash 21 multiplayer card game
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.cardclash21.com
    description: Production server
paths:
  /api/session/create:
    post:
      summary: Create new player session
      description: Creates a temporary session for a player with their display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - playerName
              properties:
                playerName:
                  type: string
                  minLength: 2
                  maxLength: 20
                  example: "Jerm"
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Invalid player name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/matchmaking/join:
    post:
      summary: Join matchmaking queue
      description: Add player to matchmaking queue to find opponent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sessionId
              properties:
                sessionId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Successfully joined queue
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [queued]
                  queuePosition:
                    type: integer
                  estimatedWaitTime:
                    type: integer
                    description: Estimated wait time in seconds
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/matchmaking/leave:
    post:
      summary: Leave matchmaking queue
      description: Remove player from matchmaking queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sessionId
              properties:
                sessionId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Successfully left queue
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [removed]
...
components:
  schemas:
    SessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        playerName:
          type: string
          example: "Jerm"
        status:
          type: string
          enum: [idle, queued, in-game]
          example: "idle"
        createdAt:
          type: integer
          description: Unix timestamp
          example: 1672531200000
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: "INVALID_SESSION"
        message:
          type: string
          example: "Session not found or expired"
        timestamp:
          type: integer

# WebSocket Events Documentation
...